Add libcrypto and new crypto.h and crypto.c files to the build system. Add sql columns to support password hashing. Finish pbkdf2 implementation. Now needs only testing.

+++ src/login/login.c
@@ -1936,7 +1936,7 @@ int login_config_read(const char* cfgName)
 // Function called at exit of the server
 //--------------------------------------
 int do_final(void) {
-	size_t i;
+	int i;
 	struct client_hash_node *hn = login_config.client_hash_nodes;
 
 	ShowStatus("Terminating...\n");
@@ -2003,7 +2003,7 @@ void do_shutdown_login(void)
 {
 	if( runflag != LOGINSERVER_ST_SHUTDOWN )
 	{
-		size_t id;
+		int id;
 		runflag = LOGINSERVER_ST_SHUTDOWN;
 		ShowStatus("Shutting down...\n");
 		// TODO proper shutdown procedure; kick all characters, wait for acks, ...  [FlavioJS]
@@ -2057,7 +2057,7 @@ void cmdline_args_init_local(void)
 //------------------------------
 int do_init(int argc, char** argv)
 {
-	size_t i;
+	int i;
 
 	// initialize engine (to accept config settings)
 	account_engine[0].db = account_engine[0].constructor();
